postgres_docker/requirements.txt
Django==3.0.7
psycopg2-binary==2.8.5

djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1

djangorestframework-simplejwt==4.8.0
simplejwt==2.0.1

django-filter==2.4.0
flake8==4.0.1
